diff cell with view mode: inline and split
