--- decidim-accountability/app/assets/stylesheets/decidim/accountability/accountability/_versions.scss
@@ -1,9 +1,0 @@
-.diff{
-  .addition{
-    background-color: #e6ffed;
-  }
-
-  .removal{
-    background-color: #ffeef0;
-  }
-}
