+++ decidim-accountability/app/views/decidim/accountability/versions/show.html.erb
@@ -9,10 +9,7 @@
     <div class="row">
       <div class="mediumlarge-8 columns">
         <div class="section">
-          <% diff_renderer.diff.each_value do |data| %>
-            <h3 class="section-heading"><%= data[:label] %></h3>
-            <%= render_diff_data(data) %>
-          <% end %>
+          <%= cell("decidim/diff", diff_renderer.diff.values) %>
         </div>
       </div>
