+++ decidim-core/app/assets/stylesheets/decidim/_variables.scss
@@ -20,6 +20,10 @@ $twitter: #55acee !default;
 $facebook: #3b5998 !default;
 $google: #dd4b39 !default;
 
+// background colors used in public diff views
+$color-addition: #e6ffed !default;
+$color-removal: #ffeef0 !default;
+
 // Define custom properties as css color variables
 
 :root{
