+++ decidim-core/app/assets/stylesheets/decidim/modules/_modules.scss
@@ -48,6 +48,7 @@
 @import "hover-section";
 @import "list-request";
 @import "floating-helper";
+@import "versions";
 
 //Process elements
 @import "process-header";
