+++ decidim-core/app/assets/stylesheets/decidim/modules/_order-by.scss
@@ -2,6 +2,7 @@
 $order-border: $border;
 
 /* Order by styles */
+.diff-view-by,
 .order-by{
   margin-bottom: 1rem;
   display: flex;
@@ -19,6 +20,7 @@ $order-border: $border;
   }
 }
 
+.diff-view-by__dropdown,
 .order-by__dropdown{
   display: flex;
   flex-wrap: wrap;
