+++ decidim-core/app/cells/decidim/diff/diff_unified.erb
@@ -1,0 +1,7 @@
+<div class="mediumlarge-12 columns">
+  <div class="card">
+    <div class="diff-data">
+      <%= output_unified_diff(data) %>
+    </div>
+  </div>
+</div>
