+++ decidim-core/app/cells/decidim/diff/show.erb
@@ -1,0 +1,5 @@
+<%= render :diff_mode_dropdown %>
+
+<% model.each do |data| %>
+  <%= attribute(data) %>
+<% end %>
